Cache slashing protection tests data
