+++ .gitignore
@@ -25,6 +25,7 @@ validators
 .vscode
 
 # Tests artifacts
+packages/validator/spec-tests
 packages/lodestar/test-logs/
 packages/beacon-state-transition/test-cache
 packages/*/benchmark_data
