+++ packages/validator/package.json
@@ -29,8 +29,8 @@
     "pretest": "yarn run check-types",
     "test:unit": "nyc --cache-dir .nyc_output/.cache -e .ts mocha 'test/unit/**/*.test.ts'",
     "test": "yarn test:unit",
-    "test:e2e": "yarn run download-spec-tests && mocha 'test/e2e/**/*.test.ts'",
-    "download-spec-tests": "node -r ts-node/register test/e2e/slashing-protection-interchange-tests/downloadTests.ts",
+    "test:e2e": "yarn run download-spec-tests && mocha 'test/spec/**/*.test.ts'",
+    "download-spec-tests": "node -r ts-node/register test/spec/downloadTests.ts",
     "coverage": "codecov -F lodestar-validator",
     "check-readme": "typescript-docs-verifier"
   },
