+++ packages/validator/test/spec/index.test.ts
@@ -10,7 +10,7 @@ import {
   SlashingProtectionAttestation,
   InvalidBlockError,
   InvalidAttestationError,
-} from "../../../src/slashingProtection";
+} from "../../src/slashingProtection";
 
 chai.use(chaiAsPromised);
